Updating the lighting interfaces.
