+++ src/light/area_light.h
@@ -17,6 +17,9 @@
 
 #include "light_interface.h"
 
+#include "../utils/image.h"
+#include "../geometry/triangle.h"
+
 namespace spica {
 
     class AreaLight : public ILight {
@@ -25,11 +28,30 @@ namespace spica {
     };
 
     class DiffuseAreaLight : public AreaLight {
-    
+    private:
+        Color _emittance;
+        Triangle _triangle;
+
+    public:
+        DiffuseAreaLight(const Triangle& tri, const Color& emittance);
+        ~DiffuseAreaLight();
+
+        Color L(const Vector3D& pos, const Vector3D& normal, const Vector3D& w) const override;
     };
 
     class InfiniteAreaLight : public AreaLight {
-    
+    private:
+        const Color& _lightness;
+        const Image& _envmap;
+
+    public:
+        InfiniteAreaLight(const Color& lightness, const Image& envmap);
+        ~InfiniteAreaLight();
+
+        Color L(const Vector3D& pos, const Vector3D& normal, const Vector3D& w) const override;
+
+    private:
+        void computeSamplePdf();
     };
 
 }  // namespace spica
