+++ src/renderer/CMakeLists.txt
@@ -8,7 +8,6 @@ set(SOURCES
     ${CMAKE_CURRENT_LIST_DIR}/bsdf.cc
     ${CMAKE_CURRENT_LIST_DIR}/brdf.cc
     ${CMAKE_CURRENT_LIST_DIR}/bssrdf.cc
-    ${CMAKE_CURRENT_LIST_DIR}/envmap.cc
     ${CMAKE_CURRENT_LIST_DIR}/renderer_helper.cc
     ${CMAKE_CURRENT_LIST_DIR}/pathtrace.cc
     ${CMAKE_CURRENT_LIST_DIR}/bdpt.cc
@@ -31,7 +30,6 @@ set(HEADERS
     ${CMAKE_CURRENT_LIST_DIR}/bsdf.h
     ${CMAKE_CURRENT_LIST_DIR}/brdf.h
     ${CMAKE_CURRENT_LIST_DIR}/bssrdf.h
-    ${CMAKE_CURRENT_LIST_DIR}/envmap.h
     ${CMAKE_CURRENT_LIST_DIR}/renderer_helper.h
     ${CMAKE_CURRENT_LIST_DIR}/pathtrace.h
     ${CMAKE_CURRENT_LIST_DIR}/bdpt.h
