+++ test/CMakeLists.txt
@@ -17,9 +17,9 @@ if (${SPICA_BUILD_TEST} STREQUAL "ON")
       set(GTEST_LIB "libgtest.a")
       set(GTEST_MAIN_LIB "libgtest_main.a")
     endif()
-    set(GTEST_LIBRARY $ENV{GTEST_ROOT}/lib/${GTEST_LIB})
-    set(GTEST_MAIN_LIBRARY $ENV{GTEST_ROOT}/lib/${GTEST_MAIN_LIB})
-    set(GTEST_INCLUDE_DIRS $ENV{GTEST_ROOT}/include)
+    set(GTEST_LIBRARY ${GTEST_ROOT}/lib/${GTEST_LIB})
+    set(GTEST_MAIN_LIBRARY ${GTEST_ROOT}/lib/${GTEST_MAIN_LIB})
+    set(GTEST_INCLUDE_DIRS ${GTEST_ROOT}/include)
   else()
     set(GTEST_LIBRARY $ENV{GTEST_LIBRARY})
     set(GTEST_MAIN_LIBRARY $ENV{GTEST_MAIN_LIBRARY})
